+++ src/components/themed/FilledTextInput.tsx
@@ -210,6 +210,11 @@ export const FilledTextInput = React.forwardRef<FilledTextInputRef, FilledTextIn
 
   const InputComponent = numeric ? StyledNumericInput : StyledAnimatedTextInput
 
+  // HACK: Some Android devices/versions, mostly Samsung, have a bug where the
+  // text input always blurs immediately after focusing.
+  const { keyboardType } = props
+  const hackKeyboardType = isAndroid && (keyboardType == null || keyboardType === 'default') ? 'visible-password' : keyboardType
+
   return (
     <View style={spaceStyle}>
       <TouchableWithoutFeedback accessible={false} testID={testID} onPress={() => focus()}>
@@ -230,7 +235,7 @@ export const FilledTextInput = React.forwardRef<FilledTextInputRef, FilledTextIn
               accessible
               animated
               ref={inputRef}
-              keyboardType={props.keyboardType}
+              keyboardType={hackKeyboardType}
               returnKeyType={props.returnKeyType}
               accessibilityState={{ disabled }}
               autoFocus={autoFocus}