+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## 4.0.3 (2024-02-15)
 
 - changed: (FIO) Use a backup balance method for accounts affected by unstake chain data issue
+- fixed: Some text inputs not selectable for some Samsung devices
 
 ## 4.0.2 (2024-02-13)